+++ lib/hcloud/concerns/deletable.rb
@@ -9,12 +9,17 @@ module HCloud
       def delete
         raise Errors::MissingIDError unless id
 
-        client
+        response = client
           .delete("#{resource_path}/#{id}")
 
         @deleted = true
 
-        self
+        # Return the resource instance
+        return self unless response.is_a?(Hash) && response.key?(:action)
+
+        # Some resources return an action instead of the resource itself (e.g. Storage Box API)
+        # Return an Action instance
+        Action.new response[:action]
       end
 
       def deleted?