+++ spec/hcloud/concerns/deletable_spec.rb
@@ -25,5 +25,20 @@ RSpec.describe HCloud::Deletable do
 
       expect(resource.delete).to be_a described_class
     end
+
+    context "when the API return an action" do
+      it "returns the action" do
+        stub_request(:delete, "https://api.hetzner.cloud/v1/examples/#{resource.id}")
+          .to_return(body: { action: { id: 1, status: "running", command: "delete_resource" } }.to_json)
+
+        action = resource.delete
+
+        expect(action).to be_a HCloud::Action
+
+        expect(action.id).to eq 1
+        expect(action.status).to eq "running"
+        expect(action.command).to eq "delete_resource"
+      end
+    end
   end
 end